Feat: Organizando melhor erros e alterando pt para en

+++ src/modules/employee/services/EmployeeService.ts
@@ -1,5 +1,6 @@
 import { EmployeeRepository } from "../../../interfaces/repositories/EmployeeRepository";
 import { EmployeeService } from "../../../interfaces/services/EmployeeService";
+import { NotFoundError } from "../../../shared/errors/AppError";
 import { CreateEmployeeDTO } from "../dtos/CreateEmployeeDTO";
 import { UpdateEmployeeDTO } from "../dtos/UpdateEmployeeDTO";
 import { Employee } from "../models/Employee";
@@ -14,7 +15,10 @@ export class EmployeeServiceImpl implements EmployeeService{
     }
     async update(id: string, updateEmployeeDTO: UpdateEmployeeDTO): Promise<Employee> {
         const employee = await this.employeeRepository.findById(id)
-
+        if(!employee){
+            throw new NotFoundError('Employee doesnt exist')
+        }
+        
         return await this.employeeRepository.update(id,updateEmployeeDTO)
     }
     async findById(id: string): Promise<Employee> {
@@ -25,7 +29,9 @@ export class EmployeeServiceImpl implements EmployeeService{
     }
     async delete(id: string): Promise<void> {
         const employee = await this.employeeRepository.findById(id)
-
+        if(!employee){
+            throw new NotFoundError('Employee doesnt exist');
+        }
         return await this.employeeRepository.delete(id)
     }
 }
